+++ CHANGELOG.md
@@ -7,4 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Added plugin in it's initial state
+* [PR-4](https://github.com/ITK-Leantime/favorite-tasks/pull/4)
+    * Make plugin compatible with leantime 3.2.0
+    * Change import from Event to EventDispatcher
+    * Remove empty `class` from template
+    * Update session notation
+    * Change dbcore import path
+
+* [PR-3](https://github.com/ITK-Leantime/favorite-tasks/pull/3)
+    * Change changefavorite to change_favorite
+
+* [PR-2](https://github.com/ITK-Leantime/favorite-tasks/pull/2)
+    * Remove `vendor-plugin.json` and `composer-plugin.json`
+
+* [PR-1](https://github.com/ITK-Leantime/favorite-tasks/pull/1)
+    * Added plugin in it's initial state